does this split work?
